Implement the WeakReference memory cache in common code.
