--- coil-core/src/jsMain/kotlin/coil3/memory/WeakMemoryCache.kt
@@ -1,3 +1,0 @@
-package coil3.memory
-
-internal actual fun RealWeakMemoryCache(): WeakMemoryCache = EmptyWeakMemoryCache()
