--- coil-core/src/nonJsCommonMain/kotlin/coil3/util/utils.nonJs.kt
@@ -1,11 +1,0 @@
-package coil3.util
-
-import kotlin.experimental.ExperimentalNativeApi
-
-@ExperimentalNativeApi // This must be propagated from the underlying native implementation.
-internal expect class WeakReference<T : Any>(referred: T) {
-    fun get(): T?
-    fun clear()
-}
-
-internal expect fun Any.identityHashCode(): Int
