--- coil-core/src/wasmJsMain
@@ -1,1 +1,0 @@
-jsMain
